+++ src/zbot/navbar/navbar.tsx
@@ -10,7 +10,7 @@ const NavBar = ({ href }: { href: string }) => {
   return (
     <nav className="fixed top-0 left-0 right-0 z-50 grid-m py-4 pointer-events-none">
       <div className="col-span-full grid grid-cols-subgrid">
-        <Logotype atTop={atTop} isMenuOpen={mobileShouldOpenBurger} href={href} />
+        {/* <Logotype atTop={atTop} isMenuOpen={mobileShouldOpenBurger} href={href} /> */}
         <BurgerOpenButton
           className="-col-end-1 place-self-end pointer-events-auto"
           onClick={() => setMobileShouldOpenBurger(true)}
